removed redundant classes and put them in one class
