+++ json_data/books.json
@@ -36,5 +36,24 @@
     "archived": false,
     "publisher": "Chris",
     "cover_state": "bad"
+  },
+  {
+    "id": 171,
+    "genre": [
+
+    ],
+    "author": [
+
+    ],
+    "source": [
+
+    ],
+    "label": [
+
+    ],
+    "publish_date": "2023-10-08",
+    "archived": false,
+    "publisher": "haha",
+    "cover_state": "bad"
   }
 ]
